Copy to clipboard text/plain + text/html
