+++ logicle/package.json
@@ -63,7 +63,6 @@
     "@readme/openapi-parser": "^2.6.0",
     "@tabler/icons-react": "^2.46.0",
     "@tailwindcss/typography": "0.5.9",
-    "@types/codemirror": "^5.60.15",
     "@uiw/react-codemirror": "^4.23.6",
     "ai": "^4.2.0",
     "ajv": "^8.17.1",
@@ -103,7 +102,7 @@
     "react-circular-progressbar": "^2.1.0",
     "react-content-loader": "^7.0.0",
     "react-day-picker": "^8.9.1",
-    "react-dom": "^18.2.0",
+    "react-dom": "^18.3.1",
     "react-hook-form": "^7.47.0",
     "react-hot-toast": "2.4.1",
     "react-i18next": "12.3.1",
@@ -134,6 +133,7 @@
     "@trivago/prettier-plugin-sort-imports": "4.2.0",
     "@types/bcryptjs": "2.4.3",
     "@types/better-sqlite3": "^7.6.9",
+    "@types/codemirror": "^5.60.15",
     "@types/i18next": "^12.1.0",
     "@types/jest": "29.5.4",
     "@types/mdast": "^4.0.4",
@@ -142,6 +142,7 @@
     "@types/nodemailer": "6.4.9",
     "@types/pg": "^8.11.2",
     "@types/react": "18.2.21",
+    "@types/react-dom": "^18.3.1",
     "@types/uuid": "9.0.3",
     "@typescript-eslint/eslint-plugin": "^8.24.0",
     "@typescript-eslint/parser": "^8.24.0",
